working on getting new viewing parties form link
